# Micro optimization of plotting