--- a/lib/matplotlib/cbook.py
+++ b/lib/matplotlib/cbook.py
@@ -1620,6 +1620,7 @@ def safe_first_element(obj):
     """
     return _safe_first_finite(obj, skip_nonfinite=False)
 
+_NoValue = object()
 
 def _safe_first_finite(obj, *, skip_nonfinite=True):
     """
@@ -1661,7 +1662,11 @@ def _safe_first_finite(obj, *, skip_nonfinite=True):
         raise RuntimeError("matplotlib does not "
                            "support generators as input")
     else:
-        return next((val for val in obj if safe_isfinite(val)), safe_first_element(obj))
+        value = next((val for val in obj if safe_isfinite(val)), _NoValue)
+        if value is _NoValue:
+            return safe_first_element(obj)
+        else:
+            return value
 
 
 def sanitize_sequence(data):