--- a/lib/matplotlib/cbook.py
+++ b/lib/matplotlib/cbook.py
@@ -1635,6 +1635,8 @@ def _safe_first_finite(obj, *, skip_nonfinite=True):
     def safe_isfinite(val):
         if val is None:
             return False
+        if isinstance(val, int):
+            return True
         try:
             return np.isfinite(val) if np.isscalar(val) else True
         except TypeError: