--- a/lib/matplotlib/cbook.py
+++ b/lib/matplotlib/cbook.py
@@ -286,7 +286,7 @@ class CallbackRegistry:
         """
         if self._signals is not None:
             _api.check_in_list(self._signals, signal=s)
-        for cid, ref in list(self.callbacks.get(s, {}).items()):
+        for ref in self.callbacks.get(s, {}).values():
             func = ref()
             if func is not None:
                 try: