--- a/lib/matplotlib/colors.py
+++ b/lib/matplotlib/colors.py
@@ -212,6 +212,7 @@ def _sanitize_extrema(ex):
 
 _nth_color_re = re.compile(r"\AC[0-9]+\Z")
 
+
 def _is_nth_color(c):
     """Return whether *c* can be interpreted as an item in the color cycle."""
     return isinstance(c, str) and _nth_color_re.match(c)