@tacaswell 2023-07-13
```python
import matplotlib
# print(matplotlib)
import pyperf

setup = """
import matplotlib.figure as mfigure

x=[1,2,3,4,5]
y=[5,6,3,3,4]
n=6
def go():
    for ii in range(n):
        fig = mfigure.Figure()
        ax = fig.subplots()
        ax.plot(x,y)
"""

runner = pyperf.Runner()
runner.timeit(name="mpl", stmt="go()", setup=setup)

```

This may be a better bench mark script.

I have concerns that the speed up is less than the std....