@tacaswell 2023-07-13
## new (this branch)
mpl: Mean +- std dev: 796 us +- 7 us

## old (3.7.2 from wheels)
mpl: Mean +- std dev: 845 us +- 13 us

using

```python
import matplotlib
# print(matplotlib)
import pyperf

setup = """
import matplotlib.figure as mfigure

x=[1,2,3,4,5]
y=[5,6,3,3,4]
n=6
fig = mfigure.Figure()
ax = fig.subplots()


def go():
    for ii in range(n):
        ax.plot(x,y)
"""

runner = pyperf.Runner()
runner.timeit(name="mpl", stmt="go()", setup=setup)

```


so I think there is a real speed up here, even if it is realitvely small against the cost of making a new `Figure` and `Axes` (that is probably why the std was so high...the first run took an order of magnitude longer!)